Fix autotest/cpp
git-svn-id: https://svn.osgeo.org/gdal/trunk@18618 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/cpp/test_cpl.cpp
@@ -176,7 +176,7 @@ namespace tut
     int sumValues(void* elt, void* user_data)
     {
         int* pnSum = (int*)user_data;
-        *pnSum += (int)elt;
+        *pnSum += (int)(long)elt;
         return TRUE;
     }
